feat(moved calendars): added timetables to own file
